% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma.pgm.R
\name{diagplot}
\alias{diagplot}
\title{Diagnostic Plots}
\usage{
diagplot(x, ...)
}
\arguments{
\item{x}{An object for which diagnostic plots should be produced.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Generic function for creating diagnostic plots.
}
