% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma.pgm.R
\name{diagplot.res.pgm.cond}
\alias{diagplot.res.pgm.cond}
\title{Diagnostic Plots for Conditional PGM Model}
\usage{
\method{diagplot}{res.pgm.cond}(x, ...)
}
\arguments{
\item{x}{An object of class \code{res.pgm.cond}.}

\item{...}{Additional arguments.}
}
\description{
Diagnostic Plots for Conditional PGM Model
}
