% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma.pgm.R
\name{plot.res.pgm.cond}
\alias{plot.res.pgm.cond}
\title{Plot Method for Conditional PGM Model}
\usage{
\method{plot}{res.pgm.cond}(x, y = NULL, level = 0.95, at = NULL, addhist = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{res.pgm.cond}.}

\item{y}{Ignored.}

\item{level}{Confidence level. Defaults to 0.95.}

\item{at}{Optional numeric vector of z values to plot.}

\item{addhist}{Logical indicating whether to add a histogram.}

\item{...}{Additional arguments.}
}
\description{
Plot Method for Conditional PGM Model
}
