% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma.pgm.R
\name{predict.res.pgm.cond}
\alias{predict.res.pgm.cond}
\title{Predict Method for Conditional PGM Model}
\usage{
\method{predict}{res.pgm.cond}(object, level = 0.95, at = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{res.pgm.cond}.}

\item{level}{Prediction interval level. Defaults to 0.95.}

\item{at}{Optional numeric vector of z values to predict at.}

\item{...}{Additional arguments.}
}
\description{
Predict Method for Conditional PGM Model
}
