% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma.pgm.R
\name{predict.res.pgm}
\alias{predict.res.pgm}
\title{Predict Method for PGM Model}
\usage{
\method{predict}{res.pgm}(object, level = 0.95, newmods = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{res.pgm}.}

\item{level}{Prediction interval level. Defaults to 0.95.}

\item{newmods}{Optional new moderator values for prediction.}

\item{...}{Additional arguments.}
}
\description{
Predict Method for PGM Model
}
