% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma.pgm.R
\name{print.summary.res.pgm.cond}
\alias{print.summary.res.pgm.cond}
\title{Print Conditional PGM Summary}
\usage{
\method{print}{summary.res.pgm.cond}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.res.pgm.cond}.}

\item{...}{Additional arguments.}
}
\description{
Print Conditional PGM Summary
}
