% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma.pgm.R
\name{print.summary.res.pgm}
\alias{print.summary.res.pgm}
\title{Print PGM Summary}
\usage{
\method{print}{summary.res.pgm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.res.pgm}.}

\item{...}{Additional arguments.}
}
\description{
Print PGM Summary
}
