% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgm_models.R
\name{summary.res.pgm.cond}
\alias{summary.res.pgm.cond}
\title{Summarize a Conditional PGM Model}
\usage{
\method{summary}{res.pgm.cond}(
  object,
  level = 0.95,
  less_than = NULL,
  newz = NULL,
  compare = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{res.pgm.cond}.}

\item{level}{Numeric scalar between 0 and 1 indicating the confidence level. Defaults to 0.95.}

\item{less_than}{Optional numeric vector to calculate P(theta < val).}

\item{newz}{Optional numeric vector of conditional z values to evaluate the summary at.}

\item{compare}{Optional numeric vector of length 2 or a two-column matrix specifying pairs of z values to compare.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
An object of class \code{summary.res.pgm.cond}.
}
\description{
Summarize a Conditional PGM Model
}
