% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma.pgm.R
\name{summary.res.pgm}
\alias{summary.res.pgm}
\title{Summarize a PGM Model}
\usage{
\method{summary}{res.pgm}(object, level = 0.95, less_than = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{res.pgm}.}

\item{level}{Confidence level. Defaults to 0.95.}

\item{less_than}{Optional numeric vector to calculate P(theta < val).}

\item{...}{Additional arguments.}
}
\description{
Summarize a PGM Model
}
