% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgm_models.R
\name{diagplot.res.pgm}
\alias{diagplot.res.pgm}
\title{Diagnostic Plots for PGM Model}
\usage{
\method{diagplot}{res.pgm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{res.pgm}.}

\item{...}{Additional graphic arguments.}
}
\description{
Diagnostic Plots for PGM Model
}
