% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma.pgm.R
\name{plot.res.pgm}
\alias{plot.res.pgm}
\title{Plot Method for PGM Model}
\usage{
\method{plot}{res.pgm}(x, y = NULL, level = 0.95, newmods = NULL, addhist = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{res.pgm}.}

\item{y}{Ignored.}

\item{level}{Confidence level. Defaults to 0.95.}

\item{newmods}{Optional new moderator values.}

\item{addhist}{Logical indicating whether to add a histogram of yi.}

\item{...}{Additional arguments.}
}
\description{
Plot Method for PGM Model
}
