% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgm_models.R
\name{predict.res.pgm.cond}
\alias{predict.res.pgm.cond}
\title{Predict Method for Conditional PGM Model}
\usage{
\method{predict}{res.pgm.cond}(object, level = 0.95, newz = NULL, digits = 4, ...)
}
\arguments{
\item{object}{An object of class \code{res.pgm.cond}.}

\item{level}{Numeric scalar indicating prediction interval level. Defaults to 0.95.}

\item{newz}{Optional numeric vector of conditional z values to predict at.}

\item{digits}{Integer indicating the number of decimal places to print. Defaults to 4.}

\item{...}{Additional arguments passed to other methods.}
}
\description{
Predict Method for Conditional PGM Model
}
