% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgm_models.R
\name{predict.res.pgm}
\alias{predict.res.pgm}
\title{Predict Method for PGM Model}
\usage{
\method{predict}{res.pgm}(object, level = 0.95, newX = NULL, digits = 4, ...)
}
\arguments{
\item{object}{An object of class \code{res.pgm}.}

\item{level}{Numeric scalar indicating the prediction interval level. Defaults to 0.95.}

\item{newX}{Optional numeric matrix of new moderator values for prediction.}

\item{digits}{Integer indicating the number of decimal places to print. Defaults to 4.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
A matrix containing the lower and upper prediction intervals.
}
\description{
Predict Method for PGM Model
}
