% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgm_models.R
\name{print.summary.res.pgm}
\alias{print.summary.res.pgm}
\title{Print PGM Summary}
\usage{
\method{print}{summary.res.pgm}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{summary.res.pgm}.}

\item{digits}{Integer indicating the number of decimal places to print. Defaults to 4.}

\item{...}{Additional arguments passed to other methods.}
}
\description{
Print PGM Summary
}
