% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma.pgm.R
\name{summary.res.pgm.cond}
\alias{summary.res.pgm.cond}
\title{Summarize a Conditional PGM Model}
\usage{
\method{summary}{res.pgm.cond}(object, level = 0.95, less_than = NULL, at = NULL, compare = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{res.pgm.cond}.}

\item{level}{Confidence level. Defaults to 0.95.}

\item{less_than}{Optional numeric vector to calculate P(theta < val).}

\item{at}{Optional numeric vector of z values to evaluate the summary at.}

\item{compare}{Optional vector or matrix to compare specific z values.}

\item{...}{Additional arguments.}
}
\description{
Summarize a Conditional PGM Model
}
