% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgm_models.R
\name{summary.res.pgm}
\alias{summary.res.pgm}
\title{Summarize a PGM Model}
\usage{
\method{summary}{res.pgm}(object, level = 0.95, less_than = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{res.pgm}.}

\item{level}{Numeric scalar between 0 and 1 indicating the confidence level. Defaults to 0.95.}

\item{less_than}{Optional numeric vector to calculate the cumulative probability P(theta < val).}

\item{...}{Additional arguments passed to other methods.}
}
\value{
An object of class \code{summary.res.pgm}.
}
\description{
Summarize a PGM Model
}
